Implement storybook preview selectors
